+++ crates/wasi-common/Cargo.toml
@@ -21,8 +21,7 @@ filetime = "0.2.7"
 lazy_static = "1.4.0"
 num = { version = "0.2.0", default-features = false }
 wig = { path = "wig", version = "0.13.0" }
-wiggle = { path = "../wiggle", default-features = false }
-wiggle-runtime = { path = "../wiggle/crates/runtime" }
+wiggle = { path = "../wiggle", default-features = false, version = "0.13.0" }
 
 [target.'cfg(unix)'.dependencies]
 yanix = { path = "yanix", version = "0.13.0" }
@@ -38,4 +37,4 @@ maintenance = { status = "actively-developed" }
 [features]
 # Need to make the wiggle_metadata feature available to consumers of this
 # crate if they want the snapshots to have metadata available.
-wiggle_metadata = ["wiggle/wiggle_metadata", "wiggle-runtime/wiggle_metadata"]
+wiggle_metadata = ["wiggle/wiggle_metadata"]
