+++ crates/wiggle/crates/generate/src/names.rs
@@ -26,7 +26,7 @@ impl Names {
     }
     pub fn builtin_type(&self, b: BuiltinType, lifetime: TokenStream) -> TokenStream {
         match b {
-            BuiltinType::String => quote!(wiggle_runtime::GuestPtr<#lifetime, str>),
+            BuiltinType::String => quote!(wiggle::GuestPtr<#lifetime, str>),
             BuiltinType::U8 => quote!(u8),
             BuiltinType::U16 => quote!(u16),
             BuiltinType::U32 => quote!(u32),
@@ -64,11 +64,11 @@ impl Names {
                 witx::Type::Builtin(builtin) => self.builtin_type(*builtin, lifetime.clone()),
                 witx::Type::Pointer(pointee) | witx::Type::ConstPointer(pointee) => {
                     let pointee_type = self.type_ref(&pointee, lifetime.clone());
-                    quote!(wiggle_runtime::GuestPtr<#lifetime, #pointee_type>)
+                    quote!(wiggle::GuestPtr<#lifetime, #pointee_type>)
                 }
                 witx::Type::Array(pointee) => {
                     let pointee_type = self.type_ref(&pointee, lifetime.clone());
-                    quote!(wiggle_runtime::GuestPtr<#lifetime, [#pointee_type]>)
+                    quote!(wiggle::GuestPtr<#lifetime, [#pointee_type]>)
                 }
                 _ => unimplemented!("anonymous type ref {:?}", tref),
             },
