+++ crates/wiggle/crates/generate/src/types/flags.rs
@@ -116,10 +116,10 @@ pub(super) fn define_flags(names: &Names, name: &witx::Id, f: &witx::FlagsDataty
         }
 
         impl ::std::convert::TryFrom<#repr> for #ident {
-            type Error = wiggle_runtime::GuestError;
-            fn try_from(value: #repr) -> Result<Self, wiggle_runtime::GuestError> {
+            type Error = wiggle::GuestError;
+            fn try_from(value: #repr) -> Result<Self, wiggle::GuestError> {
                 if #repr::from(!#ident::all()) & value != 0 {
-                    Err(wiggle_runtime::GuestError::InvalidFlagValue(stringify!(#ident)))
+                    Err(wiggle::GuestError::InvalidFlagValue(stringify!(#ident)))
                 } else {
                     Ok(#ident(value))
                 }
@@ -127,8 +127,8 @@ pub(super) fn define_flags(names: &Names, name: &witx::Id, f: &witx::FlagsDataty
         }
 
         impl ::std::convert::TryFrom<#abi_repr> for #ident {
-            type Error = wiggle_runtime::GuestError;
-            fn try_from(value: #abi_repr) -> Result<#ident, wiggle_runtime::GuestError> {
+            type Error = wiggle::GuestError;
+            fn try_from(value: #abi_repr) -> Result<#ident, wiggle::GuestError> {
                 #ident::try_from(value as #repr)
             }
         }
@@ -145,7 +145,7 @@ pub(super) fn define_flags(names: &Names, name: &witx::Id, f: &witx::FlagsDataty
             }
         }
 
-        impl<'a> wiggle_runtime::GuestType<'a> for #ident {
+        impl<'a> wiggle::GuestType<'a> for #ident {
             fn guest_size() -> u32 {
                 #repr::guest_size()
             }
@@ -154,21 +154,21 @@ pub(super) fn define_flags(names: &Names, name: &witx::Id, f: &witx::FlagsDataty
                 #repr::guest_align()
             }
 
-            fn read(location: &wiggle_runtime::GuestPtr<#ident>) -> Result<#ident, wiggle_runtime::GuestError> {
+            fn read(location: &wiggle::GuestPtr<#ident>) -> Result<#ident, wiggle::GuestError> {
                 use std::convert::TryFrom;
                 let reprval = #repr::read(&location.cast())?;
                 let value = #ident::try_from(reprval)?;
                 Ok(value)
             }
 
-            fn write(location: &wiggle_runtime::GuestPtr<'_, #ident>, val: Self) -> Result<(), wiggle_runtime::GuestError> {
+            fn write(location: &wiggle::GuestPtr<'_, #ident>, val: Self) -> Result<(), wiggle::GuestError> {
                 let val: #repr = #repr::from(val);
                 #repr::write(&location.cast(), val)
             }
         }
-        unsafe impl <'a> wiggle_runtime::GuestTypeTransparent<'a> for #ident {
+        unsafe impl <'a> wiggle::GuestTypeTransparent<'a> for #ident {
             #[inline]
-            fn validate(location: *mut #ident) -> Result<(), wiggle_runtime::GuestError> {
+            fn validate(location: *mut #ident) -> Result<(), wiggle::GuestError> {
                 use std::convert::TryFrom;
                 // Validate value in memory using #ident::try_from(reprval)
                 let reprval = unsafe { (location as *mut #repr).read() };
