+++ crates/wiggle/crates/generate/src/types/int.rs
@@ -37,15 +37,15 @@ pub(super) fn define_int(names: &Names, name: &witx::Id, i: &witx::IntDatatype) 
         }
 
         impl ::std::convert::TryFrom<#repr> for #ident {
-            type Error = wiggle_runtime::GuestError;
-            fn try_from(value: #repr) -> Result<Self, wiggle_runtime::GuestError> {
+            type Error = wiggle::GuestError;
+            fn try_from(value: #repr) -> Result<Self, wiggle::GuestError> {
                 Ok(#ident(value))
             }
         }
 
         impl ::std::convert::TryFrom<#abi_repr> for #ident {
-            type Error = wiggle_runtime::GuestError;
-            fn try_from(value: #abi_repr) -> Result<#ident, wiggle_runtime::GuestError> {
+            type Error = wiggle::GuestError;
+            fn try_from(value: #abi_repr) -> Result<#ident, wiggle::GuestError> {
                 #ident::try_from(value as #repr)
             }
         }
@@ -62,7 +62,7 @@ pub(super) fn define_int(names: &Names, name: &witx::Id, i: &witx::IntDatatype) 
             }
         }
 
-        impl<'a> wiggle_runtime::GuestType<'a> for #ident {
+        impl<'a> wiggle::GuestType<'a> for #ident {
             fn guest_size() -> u32 {
                 #repr::guest_size()
             }
@@ -71,19 +71,19 @@ pub(super) fn define_int(names: &Names, name: &witx::Id, i: &witx::IntDatatype) 
                 #repr::guest_align()
             }
 
-            fn read(location: &wiggle_runtime::GuestPtr<'a, #ident>) -> Result<#ident, wiggle_runtime::GuestError> {
+            fn read(location: &wiggle::GuestPtr<'a, #ident>) -> Result<#ident, wiggle::GuestError> {
                 Ok(#ident(#repr::read(&location.cast())?))
 
             }
 
-            fn write(location: &wiggle_runtime::GuestPtr<'_, #ident>, val: Self) -> Result<(), wiggle_runtime::GuestError> {
+            fn write(location: &wiggle::GuestPtr<'_, #ident>, val: Self) -> Result<(), wiggle::GuestError> {
                 #repr::write(&location.cast(), val.0)
             }
         }
 
-        unsafe impl<'a> wiggle_runtime::GuestTypeTransparent<'a> for #ident {
+        unsafe impl<'a> wiggle::GuestTypeTransparent<'a> for #ident {
             #[inline]
-            fn validate(_location: *mut #ident) -> Result<(), wiggle_runtime::GuestError> {
+            fn validate(_location: *mut #ident) -> Result<(), wiggle::GuestError> {
                 // All bit patterns accepted
                 Ok(())
             }
