+++ crates/wiggle/crates/generate/src/types/struct.rs
@@ -23,7 +23,7 @@ pub(super) fn define_struct(
                 witx::Type::Builtin(builtin) => names.builtin_type(*builtin, quote!('a)),
                 witx::Type::Pointer(pointee) | witx::Type::ConstPointer(pointee) => {
                     let pointee_type = names.type_ref(&pointee, quote!('a));
-                    quote!(wiggle_runtime::GuestPtr<'a, #pointee_type>)
+                    quote!(wiggle::GuestPtr<'a, #pointee_type>)
                 }
                 _ => unimplemented!("other anonymous struct members"),
             },
@@ -39,20 +39,20 @@ pub(super) fn define_struct(
             witx::TypeRef::Name(nt) => {
                 let type_ = names.type_(&nt.name);
                 quote! {
-                    let #name = <#type_ as wiggle_runtime::GuestType>::read(&#location)?;
+                    let #name = <#type_ as wiggle::GuestType>::read(&#location)?;
                 }
             }
             witx::TypeRef::Value(ty) => match &**ty {
                 witx::Type::Builtin(builtin) => {
                     let type_ = names.builtin_type(*builtin, anon_lifetime());
                     quote! {
-                    let #name = <#type_ as wiggle_runtime::GuestType>::read(&#location)?;
+                    let #name = <#type_ as wiggle::GuestType>::read(&#location)?;
                     }
                 }
                 witx::Type::Pointer(pointee) | witx::Type::ConstPointer(pointee) => {
                     let pointee_type = names.type_ref(&pointee, anon_lifetime());
                     quote! {
-                        let #name = <wiggle_runtime::GuestPtr::<#pointee_type> as wiggle_runtime::GuestType>::read(&#location)?;
+                        let #name = <wiggle::GuestPtr::<#pointee_type> as wiggle::GuestType>::read(&#location)?;
                     }
                 }
                 _ => unimplemented!("other anonymous struct members"),
@@ -64,7 +64,7 @@ pub(super) fn define_struct(
         let name = names.struct_member(&ml.member.name);
         let offset = ml.offset as u32;
         quote! {
-            wiggle_runtime::GuestType::write(
+            wiggle::GuestType::write(
                 &location.cast::<u8>().add(#offset)?.cast(),
                 val.#name,
             )?;
@@ -91,9 +91,9 @@ pub(super) fn define_struct(
         });
 
         quote! {
-            unsafe impl<'a> wiggle_runtime::GuestTypeTransparent<'a> for #ident {
+            unsafe impl<'a> wiggle::GuestTypeTransparent<'a> for #ident {
                 #[inline]
-                fn validate(location: *mut #ident) -> Result<(), wiggle_runtime::GuestError> {
+                fn validate(location: *mut #ident) -> Result<(), wiggle::GuestError> {
                     #(#member_validate)*
                     Ok(())
                 }
@@ -109,7 +109,7 @@ pub(super) fn define_struct(
             #(#member_decls),*
         }
 
-        impl<'a> wiggle_runtime::GuestType<'a> for #ident #struct_lifetime {
+        impl<'a> wiggle::GuestType<'a> for #ident #struct_lifetime {
             fn guest_size() -> u32 {
                 #size
             }
@@ -118,12 +118,12 @@ pub(super) fn define_struct(
                 #align
             }
 
-            fn read(location: &wiggle_runtime::GuestPtr<'a, Self>) -> Result<Self, wiggle_runtime::GuestError> {
+            fn read(location: &wiggle::GuestPtr<'a, Self>) -> Result<Self, wiggle::GuestError> {
                 #(#member_reads)*
                 Ok(#ident { #(#member_names),* })
             }
 
-            fn write(location: &wiggle_runtime::GuestPtr<'_, Self>, val: Self) -> Result<(), wiggle_runtime::GuestError> {
+            fn write(location: &wiggle::GuestPtr<'_, Self>, val: Self) -> Result<(), wiggle::GuestError> {
                 #(#member_writes)*
                 Ok(())
             }
