+++ crates/wiggle/crates/macro/Cargo.toml
@@ -1,21 +1,26 @@
 [package]
-name = "wiggle-runtime"
+name = "wiggle-macro"
 version = "0.13.0"
 authors = ["Pat Hickey <phickey@fastly.com>", "Jakub Konka <kubkon@jakubkonka.com>", "Alex Crichton <alex@alexcrichton.com>"]
 edition = "2018"
 license = "Apache-2.0 WITH LLVM-exception"
-description = "Runtime components of wiggle code generator"
+description = "Wiggle code generator"
 categories = ["wasm"]
 keywords = ["webassembly", "wasm"]
 repository = "https://github.com/bytecodealliance/wasmtime"
+readme = "README.md"
 include = ["src/**/*", "LICENSE"]
 
-[dependencies]
-thiserror = "1"
-witx = { path = "../../../wasi-common/wig/WASI/tools/witx", version = "0.8.4", optional = true }
+[lib]
+proc-macro = true
 
-[badges]
-maintenance = { status = "actively-developed" }
+[dependencies]
+wiggle-generate = { path = "../generate", version = "0.13.0" }
+witx = { path = "../../../wasi-common/wig/WASI/tools/witx", version = "0.8.4" }
+syn = { version = "1.0", features = ["full"] }
+
+[dev-dependencies]
+wiggle = { path = "../.." }
 
 [features]
-wiggle_metadata = ['witx']
+wiggle_metadata = []
