+++ crates/wiggle/src/error.rs
